requirements.txt
#pip freeze > requirements.txt  comando import para o texto
absl-py==2.0.0
altair==5.2.0
astunparse==1.6.3
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
dlib @ file:///D:/bld/dlib_1697537219106/work
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.5.26
fonttools==4.47.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.60.0
h5py==3.8.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.4
mdurl==0.1.2
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
scipy==1.10.0
seaborn==0.13.0
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
